Added support for multiple ranges
All (numeric) parameters always define multiple ranges.
As a special case, for some parameters, the number of ranges may
be 1. All parameters define at least one, even if that single range
happens to have unspecified endpoints (both are None).

Each endpoint may be any expression. This includes expressions with
references to other parameters. Missing endpoints default to
None which has the usual meaning. This means that the range
inclMin=(-5,10) and inclMax=(5) defines a parameter that may be
initialized to a value x, such that -5 <= x <= 5 or 10 <= x.

Parameters can still be created with single limits as before.