Design creation ignores initialization order
The initialization order turns out to be irrelevant when initializing
parameters during design creation. This should be pretty obvious, since
dependent parameters are initialized recursively anyway.

Taking initialization order into account might offer a small
performance saving, but it is unlikely to be significant. However, it's
likely that it would interfere when dealing with choices in the future.